requirements.txt
black
fastapi
mypy
pylint
python-ffmpeg
python-ffmpeg-video-streaming
uvicorn
